Fix syntax error in Google analytics snippet (fixes #10)
All pages were throwing a syntax error due to this templating using the
`'` character, which is also used in the site title (`CALM: Children's
Cancer and Leukaemia Movement`).

There might be a better way to do this but changing this to use `"`
should be sufficient to fix this, since we don't/shouldn't use this in
any page titles; might as well also make the same change to prevent a
similar issue happening for URLs.